Docker start configuration
Added SQL and sh scripts for table creation and for insert the first
user.
In the .env file a ADMIN_USER variable is added. When the postgres
docker container starts, the init scripts are executed and the user is
created as an admin.
Removed mapping.xml unused file. Changed persistence.xml configuration
that creates initial database structure, now the database is created
when the docker container starts.